requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
django-cors-headers>=3.12.0,<3.13
psycopg2>=2.9.3,<2.10
djangorestframework-simplejwt>=5.2.0,<5.3

